SAK-62 updated HomePage
